Create separated components for adc read
