+++ components/analogReader/include/taskAdc.h
@@ -1,0 +1,19 @@
+#ifndef TASKADC_H
+#define TASKADC_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+#include "esp_system.h"
+#include "esp_spi_flash.h"
+#include "driver/gpio.h"
+#include "driver/adc.h"
+#include "esp_log.h"
+
+void configAdcTask();
+static void task_adc();
+
+
+#endif
