+++ main/CMakeLists.txt
@@ -1,2 +1,2 @@
 idf_component_register(SRCS "main.c"
-                    INCLUDE_DIRS "")
+                        INCLUDE_DIRS "components")
